backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "opencv-python>=4.8.0",
    "moviepy>=1.0.3",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "scikit-image>=0.21.0",
    "yt-dlp>=2025.9.26",
    "python-dotenv>=1.1.1",
    "google-adk>=1.15.1",
]
